Pull CSS updates, dutch localization (#1)
* Style changes to pivx-style and Dutch localization.

Style changes to support the PIVX layout this includes media-queries to support multiple screen, added the Dutch localization as well.

* Style change to support the selector + added the jquery script

Addition of the css rule for the selector when it is selected it will turn green. I've also added the jquery script for the selector this will change the background colour of the parent div.